user/pix: add magic to serve XO icons with the right colours
We _could_ have served SVGs directly, but FF doesn't render
SVGs in IMG tags. So we create the SVG and render 2 PNGs
from it.

send_file() here will detect that we're serving a PNG
and set the content-type correctly. Hopefully no
buggy browsers remain that still look at the "extension"
in the url.