user/dsbackup: We now list the latest snapshot, and serve files
So the latest snapshot is listed correctly, and the files too.

One nice detail is that we spoof the name of the 'file' being
downloaded so that the 'download complete' dialog on Browse.xo
reports the title of your activity rather than 'dsbackup.php'.

Which in turn exposes a bug in Browse.xo when it comes to files
with non-ascii names. Ah, well...