user/pix: fix the invokation of rsvg-convert - use exec() instead of system()
There is no simple way to invoke a system cmd that returns the exit code and
does not print the output immediately. Strange but true.

So we abuse exec() a bit. Except that it doesn't really return the exit code
except as a pass-by-ref variable. The actual return code is the last line
of output. Useful.

In any case, with this patch the code behaves correctly if your rsvg-convert
program was having trouble.