dsbackuplib: Add support for datastore 'v1' format
This patch reworks dsbackup significantly to handle
both the old 'v0' DS format (as used by 0.82) and the
new 'v1' format, used by 0.84

A single user can have mixed backups -- as long as
each specific snapshot is of a single format, the
usage is transparent.

While we're at it, we also cleanup some tmpfiles.