admin/roles: enforce role:unassignself caps check for SYSTEM level RAs
role:unassignself is a fairly tricky capability. Is it about the specific
role that has it? Or about any role at the context (and below) where the
role is assigned?

With this patch at least we ensure users cannot shoot themselves in the foot
unassigning their coursecreator role sitewide (combined with removal
of unassignself caps)