init: Cleanup 'version' mangling
local upgrades have messed up moodle's internal 'version' variable.

We know the expectd good value of version, and the possible bad values
so a bit of simple SQL can fix this.

The 'version' value must be fixed before we call cliupdate -- hence the
direct SQL manipulation. Cleaning it up in the local/db/upgrade.php is
too late in the upgrade processing.